Refactored getTrades for MVC
